% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiTxDb.Mm.mm10.R
\docType{package}
\name{EpiTxDb.Mm.mm10}
\alias{EpiTxDb.Mm.mm10}
\alias{EpiTxDb.Mm.mm10.RMBase}
\alias{EpiTxDb.Mm.mm10.tRNAdb}
\title{Annotation package for \code{EpiTxDb} objects}
\usage{
EpiTxDb.Mm.mm10.RMBase(version = "1")

EpiTxDb.Mm.mm10.tRNAdb(version = "1")
}
\arguments{
\item{version}{a \code{character} value defining a version. Versions
available: \code{"1"}.(default: \code{version = "1"})}
}
\value{
a \code{\link[EpiTxDb:EpiTxDb-class]{EpiTxDb}} object
}
\description{
This package loads one or more \code{EpiTxDb} objects. Such \code{EpiTxDb}
objects are an R interface to prefabricated databases contained by this
package.

The names of any objects exposed by this package indicate the origin and
resources exposed. So for example \code{EpiTxDb.Mm.mm10.tRNAdb} would be a
\code{EpiTxDb} object of Mus musculus data from tRNAdb build based on the
mm10 genome.
}
\examples{
EpiTxDb.Mm.mm10.tRNAdb()
}
\seealso{
\itemize{
\item{\code{\link[EpiTxDb:modifications]{modifications}}}
\item{\code{\link[EpiTxDb:modifications]{modificationsBy}}}
\item{\code{\link[EpiTxDb:modifications]{modifiedSeqsByTranscript}}}
}
}
\author{
Felix G M Ernst [aut]
}
